requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
tortoise-orm==0.19.3
aerich==0.7.1
asyncpg==0.28.0
pydantic==2.3.0
python-dotenv==1.0.0
